refactor: remove utils
