--- providers/utils.ts
@@ -1,1 +1,0 @@
-export const getLoginId = () => 'login.id';
